Try different solution

+++ 0x12-advanced_binary_search/search_algos.h
@@ -6,6 +6,6 @@
 #include <string.h>
 
 
-int advanced_binary(int *array, int size, int value);
+int advanced_binary(int *array, size_t size, int value);
 
 #endif
